Create Package for Mode Commands
So they aren't at top level.
