+++ src/main/java/frc/robot/commands/modes/TeleopDoNothing.java
@@ -6,24 +6,26 @@
 /* of this project.                                                      */
 /*-----------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.modes;
 
+
+import frc.robot.commands.DoNothing;
 
 import riolog.PKLogger;
 import riolog.RioLogger;
 
 
 /**
- * Provides an implementation of the DoNothing command for Auto mode.
+ * Provides an implementation of the DoNothing command for Teleop mode.
  * <p>
  * Mostly so the logging and telemetry displays are more exact.
  */
-public class AutoDoNothing extends DoNothing {
+public class TeleopDoNothing extends DoNothing {
     
     /** Our classes' logger **/
-    private static final PKLogger logger = RioLogger.getLogger(AutoDoNothing.class.getName());
+    private static final PKLogger logger = RioLogger.getLogger(TeleopDoNothing.class.getName());
 
-    public AutoDoNothing() {
+    public TeleopDoNothing() {
         logger.info("constructing {}", getName());
     }
